Moved the go-server to root
